requirements.txt
pandas~=2.3.1
openai~=1.98.0
psycopg
python-dotenv~=1.1.1
timescale-vector
instructor~=1.10.0
anthropic~=0.60.0
pydantic~=2.11.7
config~=0.5.1